# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sahaayak",
    "description": "Local-language voice agent for discovering government schemes, scholarships, and jobs, and checking eligibility for them.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sessions/{session_id}/saved-benefits": {
      "get": {
        "tags": [
          "saved benefits"
        ],
        "summary": "List Saved Benefits",
        "operationId": "list_saved_benefits_api_sessions__session_id__saved_benefits_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedBenefitOut"
                  },
                  "title": "Response List Saved Benefits Api Sessions  Session Id  Saved Benefits Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "saved benefits"
        ],
        "summary": "Save Benefit",
        "operationId": "save_benefit_api_sessions__session_id__saved_benefits_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveBenefitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedBenefitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/saved-benefits/{benefit_id}": {
      "delete": {
        "tags": [
          "saved benefits"
        ],
        "summary": "Remove Saved Benefit",
        "operationId": "remove_saved_benefit_api_sessions__session_id__saved_benefits__benefit_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "benefit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Benefit Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/tasks": {
      "get": {
        "tags": [
          "saved benefits"
        ],
        "summary": "List Application Tasks",
        "operationId": "list_application_tasks_api_sessions__session_id__tasks_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "benefit_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 160
                },
                {
                  "type": "null"
                }
              ],
              "title": "Benefit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationTaskOut"
                  },
                  "title": "Response List Application Tasks Api Sessions  Session Id  Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/tasks/{task_id}": {
      "post": {
        "tags": [
          "saved benefits"
        ],
        "summary": "Update Application Task",
        "operationId": "update_application_task_api_sessions__session_id__tasks__task_id__post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationTaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationTaskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/reminders": {
      "get": {
        "tags": [
          "saved benefits"
        ],
        "summary": "List Reminders",
        "operationId": "list_reminders_api_sessions__session_id__reminders_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReminderOut"
                  },
                  "title": "Response List Reminders Api Sessions  Session Id  Reminders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "saved benefits"
        ],
        "summary": "Create Reminder",
        "operationId": "create_reminder_api_sessions__session_id__reminders_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/reminders/{reminder_id}": {
      "delete": {
        "tags": [
          "saved benefits"
        ],
        "summary": "Cancel Reminder",
        "operationId": "cancel_reminder_api_sessions__session_id__reminders__reminder_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "reminder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reminder Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationTaskOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "benefit_id": {
            "type": "string",
            "title": "Benefit Id"
          },
          "benefit_name": {
            "type": "string",
            "title": "Benefit Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "source_revision": {
            "type": "integer",
            "title": "Source Revision"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "benefit_id",
          "benefit_name",
          "kind",
          "title",
          "description",
          "position",
          "status",
          "due_at",
          "completed_at",
          "source_revision",
          "created_at",
          "updated_at"
        ],
        "title": "ApplicationTaskOut"
      },
      "ApplicationTaskUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "skipped"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ApplicationTaskUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReminderOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "benefit_id": {
            "type": "string",
            "title": "Benefit Id"
          },
          "benefit_name": {
            "type": "string",
            "title": "Benefit Name"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "delivered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "contact_display_suffix": {
            "type": "string",
            "title": "Contact Display Suffix",
            "default": ""
          },
          "delivery_status": {
            "type": "string",
            "title": "Delivery Status",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "benefit_id",
          "benefit_name",
          "due_at",
          "note",
          "timezone",
          "channel",
          "status",
          "created_at",
          "delivered_at"
        ],
        "title": "ReminderOut"
      },
...
      "SaveBenefitRequest": {
        "properties": {
          "benefit_id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Benefit Id"
          }
        },
        "type": "object",
        "required": [
          "benefit_id"
        ],
        "title": "SaveBenefitRequest"
      },
      "SavedBenefitOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "benefit_id": {
            "type": "string",
            "title": "Benefit Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "state_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Code"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "source_title": {
            "type": "string",
            "title": "Source Title"
          },
          "source_document_url": {
            "type": "string",
            "title": "Source Document Url"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "saved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Saved At"
          },
          "job_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Job Metadata",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id",
          "benefit_id",
          "name",
          "domain",
          "state_code",
          "description",
          "source_title",
          "source_document_url",
          "verification_status",
          "saved_at"
        ],
        "title": "SavedBenefitOut"
      },
...
    }
  }
}